Add automatic CCL list generation for prefill and decode when user does not provide lists (#663)
For using CCL feature during prefilling and decoding, the user needs to
pass two lists of context lengths to be used during these processes.
Here, we add an option to generate these lists automatically when they
are not provided by the user manually. These list generation is suitable
for a general-purpose application which considers both prefilling and
decoding processes and generates CCL lists for both.

---------

+++ QEfficient/transformers/models/llava/modeling_llava.py
@@ -181,7 +181,7 @@ class QEffLlavaForConditionalGeneration(LlavaForConditionalGeneration):
         lang_inputs["position_ids"] = torch.full(lang_inputs["position_ids"].shape, CTX_LEN - 1)
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.long)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
 
         if continuous_batching:
             lang_inputs["batch_index"] = torch.arange(BS).view(BS, 1)
